requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-filter==2.4.0
django-fontawesome-5==1.0.18
django-reset-migrations==0.4.0
django-taggit==1.5.1
django-taggit-labels==0.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
munch==2.5.0
mysql-connector-python==8.0.25
Pillow==8.3.2
pip-search==0.0.7
protobuf==3.17.3
psycopg2==2.9.1
Pygments==2.9.0
PyMySQL==1.0.2
pypisearch==1.3.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rich==10.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
urllib3==1.26.6
whitenoise==5.2.0
